workers/WorkerGlobalScope_ErrorEvent_message.htm

<!DOCTYPE html>
<title> WorkerGlobalScope onerror event handler argument: message </title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
async_test(function(t) {
  var worker = new Worker('./support/ErrorEvent.js');
  worker.onmessage = t.step_func_done(function(e) {
    assert_greater_than(e.data.message.indexOf("Error Message"), -1);
  });
  worker.postMessage("Error Message");
});
</script>
